make UIS non-kryo serializable
